+++ src/controllers/Game.js
@@ -128,7 +128,10 @@ exports.getFindSearch = (query, req) => {
 
 exports.getFindProjection = query => query.fields ? query.fields.split(",") : null;
 
-exports.getFindOptions = options => ({ limit: options.limit });
+exports.getFindOptions = options => ({
+    limit: options.limit,
+    sort: { [options["sort-by"]]: options.order },
+});
 
 exports.getGames = async(req, res) => {
     try {